Update NNW_Structure.py
Foward Propagation comentaries were corrected. Now the coments say that you can use an input with as many training samples as you like, so, as many lines as you want.

BackPropagation function was started also, The first gradient calculation is implemented and the propagation of the error using this gradient, on the weights of the hidden layer was also implemented.